chore(build): cleanup for rollup
